Refactor import statements in UserTable and UserActions components
